% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees_meteo_station.R
\name{donnees_meteo_station}
\alias{donnees_meteo_station}
\title{Recuperer et formater les donnees d'une station meteo d'Environnement Canada}
\usage{
donnees_meteo_station(
  stationid = 30165,
  year = 2020,
  returndf = FALSE,
  savecsv = FALSE,
  force_download = FALSE
)
}
\arguments{
\item{stationid}{: permet a l'utilisateur de preciser quelle station meteo il souhaite consulter. Liste complete (fichier csv officiel d'Environnement et Changement climatique Canada ) : https://drive.google.com/file/d/1MnCvrLiCz_V8y8gCYMX_Mucerupp5UsF/view?usp=sharing
Par defaut va chercher les donnees liees a la station de l'aeroport de Montreal (30165).}

\item{year}{: permet de choisir l'annee de telechargement, 2020 par defaut}

\item{returndf}{: precise si la fonction enregistre le data frame contenant les donnees dans l'environnement global (par defaut) ou s'il retourne un data frame avec retrun().}

\item{savecsv}{: permet d'enregistrer les donnees recuperes, nettoyees et compilees dans un csv en local.  False par defaut.}

\item{force_download}{: pour eviter de surcharger les serveurs d'Environnement Canada, la fonction ne telecharge les donnees que si elles n'existent pas deja dans le repertoire local, possibilite de forcer le telechargement, pour mettre a jour par exemple.}
}
\description{
La fonction va aller chercher les donnees sur le serveur d'Environnement Canada, concatener les 12 fichiers mensuels,
formater et nettoyer les donnees. Puis finalement soit enregistrer directement un data frame dans l'environnement
global soit retourner un data frame pour etre travaille par l'utilisateur en aval.
}
\examples{
donnees_meteo_station()
donnees_meteo_station(stationid=27226,year=2019,savecsv=TRUE)
}
