% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphique_meteo.R
\name{graphique_meteo}
\alias{graphique_meteo}
\title{Produire des graphiques avec les donnees obtenues grace a la fonction donnees_meteo_station}
\usage{
graphique_meteo(donnees, mois = 0, graphiques = "tv")
}
\arguments{
\item{donnees}{: un data_frame contenant au moins les colonnes suivantes:
- Date_Heure, POSIXct
- Temp_C, numeric
- Vit_du_vent_km_h, numeric}

\item{mois}{: l'utilisateur peut preciser un mois en particulier pour le graphique, par defaut 0 : affiche tous les mois de l’annee}

\item{graphiques}{: "t" pour juste le graphique des temperatures, "v" pour le graphique de la vitesse du vent et par defaut "tv" pour les deux graphiques.}
}
\description{
La fonction utilise les fonctions des librairies ggplot et dplyr pour agreger les donnees par jour puis produire deux graphiques.
Lors de son initialisation elle verifie que les libraires necessaires sont bien installees, et que les donnees sont au bon format.
}
\examples{
dm <-donnees_meteo_station(114)
graphique_meteo(dm)
graphique_meteo(dm,mois=7,"t")
}
